[toml] Better String checks and refactor
