+++ code/lib/toml/parse/src/main/scala/dev/librecybernetics/parser/toml/base/String.scala
@@ -1,52 +1,12 @@
 package dev.librecybernetics.parser.toml.base
 
-import java.lang.Integer
-import scala.language.postfixOps
-
 import cats.implicits.*
 import cats.parse.Parser
-import cats.parse.Accumulator0.charStringAccumulator0
 
 import dev.librecybernetics.parser.*
 
-private val singleQuote = Parser.char('\'')
 private val doubleQuote = Parser.char('"')
-
-private val tripleSingleQuote = Parser.string("'''").withContext("tripleSingleQuote")
 private val tripleDoubleQuote = Parser.string("\"\"\"").withContext("tripleDoubleQuote")
-
-val simpleLiteral: Parser[String] =
-  Parser
-    .charsWhile(c => !(Set('\'', '\n') contains c))
-    .surroundedBy(singleQuote)
-
-val multilineLiteral: Parser[String] =
-  (
-    newline.?.with1 *> (
-      (Parser.not(tripleSingleQuote).with1 *> Parser.anyChar).rep.string |
-        (Parser.char('\'').string <* Parser.peek(tripleSingleQuote <* Parser.not(singleQuote))).backtrack
-    ).withContext("multilineLiteral.line")
-      .rep
-      .map(_.toList.mkString(""))
-  ).surroundedBy(tripleSingleQuote).withContext("multilineLiteral")
-
-private val escaped: Parser[Char] =
-  Map(
-    "\\b"  -> '\b',
-    "\\t"  -> '\t',
-    "\\n"  -> '\n',
-    "\\f"  -> '\f',
-    "\\r"  -> '\r',
-    "\\\"" -> '"',
-    "\\\\" -> '\\'
-  ).map { (s, c) =>
-    Parser.string(s).map(_ => c)
-  }.reduceOption(_.backtrack | _.backtrack)
-    .getOrElse(Parser.fail)
-
-private val escapedNewline: Parser[String] =
-  (Parser.string("\\\n").map(_ => "") <* emptyLine.backtrack.rep0 <* spaces)
-    .withContext("escapedNewline")
 
 val simpleString: Parser[String] =
   (
@@ -62,10 +22,17 @@ val multilineString: Parser[String] =
       (Parser.not(tripleDoubleQuote | backslash).with1 *> Parser.anyChar).rep.string |
         escaped.backtrack | escapedUnicode4.backtrack | escapedUnicode8.backtrack | escapedNewline.backtrack |
         (Parser.char('\"').string <* Parser.peek(tripleDoubleQuote <* Parser.not(doubleQuote))).backtrack
-    ).withContext("multilineString.line").rep.map(_.toList.mkString(""))
-  ).surroundedBy(tripleDoubleQuote).withContext("multilineString")
+    )
+      .withContext("multilineString.line")
+      .rep
+      .map(_.toList.mkString(""))
+  )
+    .surroundedBy(tripleDoubleQuote)
+    .withContext("multilineString")
 
 val string: Parser[String] =
   // Note: "" is a valid simple string, check triple quote first
-  multilineLiteral.backtrack | simpleLiteral.backtrack |
-    multilineString.backtrack | simpleString.backtrack
+  (
+    multilineLiteral.backtrack | simpleLiteral.backtrack |
+      multilineString.backtrack | simpleString.backtrack
+  ).checkDisallowedChars
